Fix tests: - Add flow fix me comment for now (addresses same issue as [here](https://github.com/Instawork/hyperview/blob/c48e093b6473237dceb516638ca98bb25d67623b/src/services/index.js#L133-L134)) - Remove circular dependency, by moving `addHref` to `hyper-ref` module
